mock: msw 데이터 유지되도록 구현
